fixed crc.c warnings
git-svn-id: https://c64preservation.com/svn/nibtools/trunk@416 048eaad1-c872-4173-8c30-523bb29f53ec

+++ crc.h
@@ -37,7 +37,7 @@ typedef unsigned short  crc;
 
 #elif defined(CRC16)
 
-typedef unsigned short  crc;
+typedef unsigned short crc;
 
 #define CRC_NAME			"CRC-16"
 #define POLYNOMIAL			0x8005
@@ -49,7 +49,7 @@ typedef unsigned short  crc;
 
 #elif defined(CRC32)
 
-typedef unsigned long  crc;
+typedef unsigned int crc;
 
 #define CRC_NAME			"CRC-32"
 #define POLYNOMIAL			0x04C11DB7
